Several adjustments to PR #456
