+++ asistente_ladm_col/ui/reports/annex_17_map_report_dialog.ui
@@ -11,7 +11,7 @@
    </rect>
   </property>
   <property name="windowTitle">
-   <string>Map Report</string>
+   <string>Annex 17 Map Report</string>
   </property>
   <layout class="QGridLayout" name="gridLayout">
    <item row="1" column="0" colspan="3">
